Added support for sending policy ACLs to the GSA. This change depends on Connector Manager change r2883 to set the googleFeedHost property.
Added NotesConnector setters for googleConnectorName (used in the
policy ACL); policyAclPattern (used to create the policy ACL);
googleFeedHost (set by the connector manager); gsaUsername and
gsaPassword (set on the connector configuration form).

Modified NotesDatabasePoller to send an updated policy ACL to the
GSA when the database ACL has changed

Added check and workaround for null usernames in
NotesAuthorizationManager; this is a potential GSA issue.

Added synchronization around uses of the user/group cache.

Added getCommonUserName to the client API.

Modified NotesConnectorType to add the GSA username/password
values to the form and to retrieve form text from a properties
file.

Added group resolution to NotesAuthenticationManager. Groups are
prefixed and URL encoded.

Added NCCONST.VIEWPEOPLENAMELOOKUP (and the corresponding view in
the connector database template) because of an undiagnosed
problem with an equivalent view in the template.

Refactored client tests to use a shared base class for test
setup. Added code in client tests to force the initialization of
the user/group cache.

[Checked in without completed code review.]