Updated version to M7.
Exposed maintenance thread and crawler thread setup in
NotesConnector and used that in tests to avoid extra thread
creation.

NotesCrawlerThread.java: Modified getDocumentReaderNames to avoid
creating a readers list consisting of only authors. There must be
at least one value in a Readers item before Authors item values
will be preserved in a readers field. We're currently
using the DOCAUTHORREADERS field as the list of allowed readers for
authorization purposes. The DOCREADERS field is also populated
because the template's security view uses this field.

NotesConnectorDocumentList.java: when the RetainMetaData flag is
set to "no" and there are no values in the DOCAUTHORREADERS
field, the document records for indexed documents are deleted.

http://codereview.appspot.com/5561043/