First draft of a database-based UserGroupManager.
A new class, NotesUserGroupManagerDatabase, has been added. This
class creates and maintains the user and group cache in the
connector manager-provided database. This class also creates and
maintains a role cache using the database ACL information from
the configured crawl databases.

NotesUserGroupManagerDatabase is not currently used by the
connector. Upcoming tasks will replace the current
UserGroupManager with this version and integrate it with the
connector.


Remaining tasks include:

* Add methods for the connector to use to retrieve user/group
  data. Modify the rest of the connector to use these methods.

* Move the SQL statements to a resource file.

* Test with the other supported connector manager databases;
  update SQL as needed.



http://codereview.appspot.com/6062046/