Integrate the database-based UserGroupManager with the rest of the connector.
The original NotesUserGroupManager has been replaced with the
database-based implementation. The uses of the UserGroupManager
in NotesConnectorDocument, NotesAuthenticationManager,
NotesAuthorizationManager, and NotesDatabasePoller have been
updated. The constants in NCCONST referring to the user and group
cache views in the connector database have been removed.

NotesDatabasePoller no longer caches role information since
that's done in the UserGroupManager.


http://codereview.appspot.com/6220056/