Remove obsolete database ACL code that conflicts with r413.
The initial new connector code in r211 did not support policy ACLs,
inherited ACLs, or readers. It simply copied the database ACL for each
document. As part of that, if the database ACL changed, all documents
in that source database were reindexed to pick up the new database
ACL. This code was not removed when support for policy ACLs and later
inherited ACLs were added, but it serves no purpose. We never copy the
database ACL.

This code conflicts with the recent change to searching by lastUpdated
in r413, but I was also removing it along with support for policy ACLs
for version 3.4. I've split this off separately to fix the conflict
with r413 for version 3.2.10.

Code review: https://codereview.appspot.com/185180043/